+++ src/app/billing/Dictionary/nutrient/nutrient.component.ts
@@ -5,6 +5,7 @@ import { LogicNutrientComponent } from './logic-nutrient/logic-nutrient.componen
 import { FormGroup, FormBuilder } from '@angular/forms';
 import { DictionaryService } from '../dictionary.service';
 import { DeleteConfirmDailogComponent } from 'src/app/shared/delete-confirm-dailog/delete-confirm-dailog.component';
+import { RestrictionComponent } from 'src/app/shared/restriction/restriction.component';
 
 @Component({
   selector: 'app-nutrient',
@@ -13,11 +14,16 @@ import { DeleteConfirmDailogComponent } from 'src/app/shared/delete-confirm-dail
 })
 export class NutrientComponent implements OnInit {
 
-  filterToggle:boolean;
+  CreatePermission: any;
+  ReadPermission: any;
+  UpdatePermission: any;
+  DeletePermission: any;
+
+  filterToggle: boolean;
   toggleFilter() {
     this.filterToggle = !this.filterToggle;
   }
-  
+
   fakedata: any;
   NutrientList: any;
   pageNutrientList: any;
@@ -38,23 +44,32 @@ export class NutrientComponent implements OnInit {
 
   constructor(private _fb: FormBuilder, public dialog: MatDialog, private alertService: AlertService,
     private dictionaryService: DictionaryService) {
-      this.filterForm = this._fb.group({
-        'keyWord': [null]
-      });
-     }
+    this.filterForm = this._fb.group({
+      'keyWord': [null]
+    });
+  }
 
   ngOnInit() {
-    this.fakedata= [
-      {id:'1', Nutrient_Condition:'Choline',  Section:'Pregnancy & Lactation' },
-      {id:'2', Nutrient_Condition:'Choline',  Section:'Risks of Metabolic imbalances' },
-      {id:'3', Nutrient_Condition:'Choline',  Section:'Physical Activity' },
-      {id:'4', Nutrient_Condition:'Choline',  Section:'Sports Performance'},
-      {id:'5', Nutrient_Condition:'Choline',  Section:'Risks of Metabolic imbalances'},
-      {id:'6', Nutrient_Condition:'Vitamin B', Section:'Sports Performance'},
-      {id:'7', Nutrient_Condition:'Vitamin B', Section:'Nutrition'},
-      {id:'8', Nutrient_Condition:'Vitamin B', Section:'Risks of Metabolic imbalances'},
-      {id:'9', Nutrient_Condition:'Vitamin B', Section:'Nutrition' },
-    ];
+
+    let getPermissions = JSON.parse(localStorage.getItem('Permissions'));
+    // console.log('Permissions: ', getPermissions);
+
+    if (getPermissions) {
+      for (let i = 0; i < getPermissions.length; i++) {
+        let ScreenName = getPermissions[i]['ScreenName']
+        if (ScreenName == 'Nurtients/Condition') {
+          this.CreatePermission = getPermissions[i]['Create']
+          this.ReadPermission = getPermissions[i]['Read']
+          this.UpdatePermission = getPermissions[i]['Update']
+          this.DeletePermission = getPermissions[i]['Delete']
+        }
+      }
+    }
+    // console.log(this.CreatePermission, 'CreatePermission');
+    // console.log(this.ReadPermission, 'ReadPermission');
+    // console.log(this.UpdatePermission, 'UpdatePermission');
+    // console.log(this.DeletePermission, 'DeletePermission');
+
     this.getNutrientList();
   }
 
@@ -78,35 +93,45 @@ export class NutrientComponent implements OnInit {
   }
 
   public openLogicDialog(id, action, item) {
-    let dialogRef = this.dialog.open(LogicNutrientComponent, {
-      height: 'auto',
-      width: '600px',
-      autoFocus: false,
-    });
-    (<LogicNutrientComponent>dialogRef.componentInstance).id = id;
-    (<LogicNutrientComponent>dialogRef.componentInstance).action = action;
-    (<LogicNutrientComponent>dialogRef.componentInstance).item = item;
-    dialogRef.afterClosed().subscribe(data => {
-      this.getNutrientList();
-    });
+    if (this.UpdatePermission == true) {
+      let dialogRef = this.dialog.open(LogicNutrientComponent, {
+        height: 'auto',
+        width: '600px',
+        autoFocus: false,
+      });
+      (<LogicNutrientComponent>dialogRef.componentInstance).id = id;
+      (<LogicNutrientComponent>dialogRef.componentInstance).action = action;
+      (<LogicNutrientComponent>dialogRef.componentInstance).item = item;
+      dialogRef.afterClosed().subscribe(data => {
+        this.getNutrientList();
+      });
+    } else {
+      this.restrictionDialog('Update');
+    }
+
   }
 
   deleteNutrient(data) {
     let body = {
       'Nutrients_ConditionsId': data
     }
-    this.dictionaryService.deletenutrients_conditions(body).subscribe(
-      data => {
-        console.log(data)
-        if (data['Success'] == true) {
-          this.getNutrientList();
-          this.alertService.createAlert('Successfully Deleted', 1);
-        } else {
-          this.alertService.createAlert('Something Went Wrong', 0);
-        }
+    if (this.DeletePermission == true) {
+      this.dictionaryService.deletenutrients_conditions(body).subscribe(
+        data => {
+          console.log(data)
+          if (data['Success'] == true) {
+            this.getNutrientList();
+            this.alertService.createAlert('Successfully Deleted', 1);
+          } else {
+            this.alertService.createAlert('Something Went Wrong', 0);
+          }
 
-      }
-    );
+        }
+      );
+    } else {
+      this.restrictionDialog('Delete');
+    }
+
   }
 
   filterBy(formValues) {
@@ -149,17 +174,28 @@ export class NutrientComponent implements OnInit {
     this.pageNutrientList = this.NutrientList.slice(this.currentPageTemp * this.pageSize, (this.currentPageTemp * this.pageSize) + this.pageSize);
   }
 
-  public deleteDialog(id){
+  public deleteDialog(id) {
     let dialogRef = this.dialog.open(DeleteConfirmDailogComponent, {
       height: 'auto',
       width: '500px',
       autoFocus: false,
     });
     dialogRef.afterClosed().subscribe(data => {
-      console.log("datr",data)
-      if (data == true){
+      console.log("datr", data)
+      if (data == true) {
         this.deleteNutrient(id)
       }
+    });
+  }
+
+  public restrictionDialog(action) {
+    let dialogRef = this.dialog.open(RestrictionComponent, {
+      height: 'auto',
+      width: '500px',
+      autoFocus: false,
+    });
+    (<RestrictionComponent>dialogRef.componentInstance).action = action;
+    dialogRef.afterClosed().subscribe(data => {
     });
   }
